Read the imageRef from fields not hardcoded
