+++ pkg/img/layer.go
@@ -4,7 +4,14 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"context"
+	"errors"
+	"github.com/google/go-containerregistry/pkg/authn"
+	"github.com/google/go-containerregistry/pkg/crane"
+	containerregistryv1 "github.com/google/go-containerregistry/pkg/v1"
+	"github.com/kyma-project/lifecycle-manager/pkg/ocmextensions"
 	apimetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"regexp"
 )
 
 type LayerName string
@@ -14,6 +21,8 @@ const (
 	CRDsLayer                LayerName = "crds"
 	AssociatedResourcesLayer LayerName = "associated-resources"
 )
+
+var ErrImageLayerPull = errors.New("failed to pull layer")
 
 type LayerRepresentation interface {
 	ToInstallRaw() ([]byte, error)
@@ -48,3 +57,24 @@ type (
 )
 
 type Layers []Layer
+
+func PullLayer(ctx context.Context, imageRef string, keyChain authn.Keychain) (containerregistryv1.Layer, error) {
+	noSchemeImageRef := ocmextensions.NoSchemeURL(imageRef)
+	isInsecureLayer, err := regexp.MatchString("^http://", imageRef)
+	if err != nil {
+		return nil, fmt.Errorf("invalid imageRef: %w", err)
+	}
+
+	if isInsecureLayer {
+		imgLayer, err := crane.PullLayer(noSchemeImageRef, crane.Insecure, crane.WithAuthFromKeychain(keyChain))
+		if err != nil {
+			return nil, fmt.Errorf("%s due to: %w", ErrImageLayerPull.Error(), err)
+		}
+		return imgLayer, nil
+	}
+	imgLayer, err := crane.PullLayer(noSchemeImageRef, crane.WithAuthFromKeychain(keyChain), crane.WithContext(ctx))
+	if err != nil {
+		return nil, fmt.Errorf("%s due to: %w", ErrImageLayerPull.Error(), err)
+	}
+	return imgLayer, nil
+}
